refactor: simplify TrainModelDialog
Combine the `TrainModelModal` and `RetrainModelModal` into one since we
don't really need the extra composition.
In both cases we say "Train model", the main difference is that we can
choose a different base modell, which then limits the selected policy to
the selected base model.
